vehicle_yolo/requirements.txt
PyYAML==5.4.1
easydict==1.9
numpy==1.24.3
ultralytics==8.3.217
tqdm>=4.66.0